Kill `other/any` task group
