+++ buildSrc/src/main/kotlin/io/spine/internal/gradle/javascript/task/impl/Assemble.kt
@@ -26,50 +26,3 @@
 
 package io.spine.internal.gradle.javascript.task.impl
 
-import io.spine.internal.gradle.javascript.task.JsTaskConfiguring
-import io.spine.internal.gradle.javascript.task.JsTaskRegistering
-import io.spine.internal.gradle.javascript.task.buildJs
-import io.spine.internal.gradle.javascript.task.testJs
-import org.gradle.api.tasks.Copy
-import org.gradle.kotlin.dsl.create
-
-fun JsTaskRegistering.webpack() {
-    copyBundledJs()
-}
-
-/**
- * Extends `buildJs` and `testJs` to run `webpack` builds.
- */
-fun JsTaskConfiguring.webPack() {
-
-    // Customizes `buildJs` task with running `webpack` bundler.
-
-    buildJs.apply {
-
-        outputs.dir(webPackOutput)
-
-        doLast {
-            npm("run", "build")
-            npm("run", "build-dev")
-        }
-    }
-
-    // Customizes `testJs` task with running JavaScript tests.
-
-    testJs.apply {
-
-        doLast {
-            npm("run", "test")
-        }
-    }
-}
-
-private fun JsTaskRegistering.copyBundledJs() =
-    create<Copy>("copyBundledJs") {
-
-        description = "Copies bundled JavaScript sources to the NPM publication directory."
-        group = jsAnyTask
-
-        from(buildJs.outputs)
-        into(webPackPublicationDir)
-    }
